stats: freq calculation overflow
The gint freq could hold the value 2^31-1, that later used in the code as:
```
timespec_add_msec(&timer->expires, freq * 1000);
```
While the timespec_add_msec last argument is a glong that could hold 1000*2^31-1,
the multiplication happens in the gint domain, which overflows yielding a
negative number. Causing the timer imidiate trigger.

+++ lib/stats/stats.c
@@ -175,7 +175,7 @@ stats_publish_and_prune_counters(StatsOptions *options)
 static void
 stats_timer_rearm(StatsOptions *options, struct iv_timer *timer)
 {
-  gint freq = options->log_freq;
+  glong freq = options->log_freq;
 
   if (!freq)
     freq = options->lifetime <= 1 ? 1 : options->lifetime / 2;
